WIP: Set overwrite to true; disable validation temporarily
