+++ pipelines/test/integration/disturbance/test_gadm_dist_alerts_by_natural_lands.py
@@ -4,21 +4,21 @@ from unittest.mock import patch
 from pandera.pandas import DataFrameSchema, Column, Check
 from prefect.testing.utilities import prefect_test_harness
 
-from pipelines.disturbance.prefect_flows import dist_alerts_by_drivers_count
+from pipelines.disturbance.prefect_flows import dist_alerts_by_natural_lands_count
 
 
 @pytest.mark.integration
 @pytest.mark.slow
 @patch("pipelines.prefect_flows.common_stages._save_parquet")
 @patch("pipelines.prefect_flows.common_stages._load_zarr")
-def test_gadm_dist_alerts_by_driver_happy_path(
+def test_gadm_dist_alerts_happy_path(
     mock_load_zarr,
     mock_save_parquet,
     dist_ds,
     country_ds,
     region_ds,
     subregion_ds,
-    dist_drivers_ds,
+    natural_lands_ds,
 ):
     """Test full workflow with in-memory dependencies"""
 
@@ -27,33 +27,34 @@ def test_gadm_dist_alerts_by_driver_happy_path(
         country_ds,
         region_ds,
         subregion_ds,
-        dist_drivers_ds,
+        natural_lands_ds,
     ]
 
     with prefect_test_harness():
-        result_uri = dist_alerts_by_drivers_count(
+        result_uri = dist_alerts_by_natural_lands_count(
             dist_zarr_uri="s3://dummy_zarr_uri",
             dist_version="test_v1",
+            overwrite=True,
         )
 
     assert (
         result_uri
-        == "s3://gfw-data-lake/umd_glad_dist_alerts/test_v1/tabular/zonal_stats/gadm/gadm_adm2_dist_alerts_by_drivers.parquet"
+        == "s3://gfw-data-lake/umd_glad_dist_alerts/test_v1/tabular/zonal_stats/gadm/gadm_adm2_dist_alerts_by_natural_lands.parquet"
     )
 
 
-@pytest.mark.slow
 @pytest.mark.integration
+@pytest.mark.slow
 @patch("pipelines.prefect_flows.common_stages._save_parquet")
 @patch("pipelines.prefect_flows.common_stages._load_zarr")
-def test_gadm_dist_alerts_by_driver_result(
+def test_gadm_dist_alerts_result(
     mock_load_zarr,
     mock_save_parquet,
     dist_ds,
     country_ds,
     region_ds,
     subregion_ds,
-    dist_drivers_ds,
+    natural_lands_ds,
 ):
     alert_schema = DataFrameSchema(
         name="GADM Dist Alerts",
@@ -61,7 +62,7 @@ def test_gadm_dist_alerts_by_driver_result(
             "country": Column(int, Check.ge(0)),
             "region": Column(int, Check.ge(0)),
             "subregion": Column(int, Check.ge(0)),
-            "driver": Column(int, Check.ge(0)),
+            "natural_lands": Column(int, Check.ge(0)),
             "alert_date": Column(
                 int,
                 checks=[
@@ -72,12 +73,19 @@ def test_gadm_dist_alerts_by_driver_result(
             "confidence": Column(int, Check.isin([2, 3])),
             "value": Column(int, Check.isin([0, 1, 2])),
         },
-        unique=["country", "region", "subregion", "driver", "alert_date", "confidence"],
+        unique=[
+            "country",
+            "region",
+            "subregion",
+            "natural_lands",
+            "alert_date",
+            "confidence",
+        ],
         checks=Check(
             lambda df: (
-                df.groupby(["country", "region", "subregion", "driver", "alert_date"])[
-                    "confidence"
-                ].transform("nunique")
+                df.groupby(
+                    ["country", "region", "subregion", "natural_lands", "alert_date"]
+                )["confidence"].transform("nunique")
                 == 1
             ),
             name="two_confidences_per_group",
@@ -90,16 +98,17 @@ def test_gadm_dist_alerts_by_driver_result(
         country_ds,
         region_ds,
         subregion_ds,
-        dist_drivers_ds,
+        natural_lands_ds,
     ]
 
     with prefect_test_harness():
-        dist_alerts_by_drivers_count(
+        dist_alerts_by_natural_lands_count(
             dist_zarr_uri="s3://dummy_zarr_uri",
             dist_version="test_v1",
+            overwrite=True,
         )
 
     # Verify
     result = mock_save_parquet.call_args[0][0]
-    print(f"\nGADM dist alerts result:\n{result}")
+    print(f"\nGADM dist alerts by natural lands result:\n{result}")
     alert_schema.validate(result)
